update id type in repository class

+++ src/main/java/com/enrique/reservatusalaback/repository/BusinessRepository.java
@@ -3,6 +3,6 @@ package com.enrique.reservatusalaback.repository;
 import com.enrique.reservatusalaback.model.Business;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BusinessRepository extends JpaRepository<Business, Integer> {
+public interface BusinessRepository extends JpaRepository<Business, Long> {
 
 }
