+++ src/main/java/com/enrique/reservatusalaback/repository/MaterialRepository.java
@@ -3,6 +3,6 @@ package com.enrique.reservatusalaback.repository;
 import com.enrique.reservatusalaback.model.Material;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MaterialRepository extends JpaRepository<Material, Integer> {
+public interface MaterialRepository extends JpaRepository<Material, Long> {
 
 }
