+++ src/main/java/com/enrique/reservatusalaback/repository/OperationRepository.java
@@ -3,6 +3,6 @@ package com.enrique.reservatusalaback.repository;
 import com.enrique.reservatusalaback.model.Operation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OperationRepository extends JpaRepository<Operation, Integer> {
+public interface OperationRepository extends JpaRepository<Operation, Long> {
 
 }
